requirements.txt
python-dotenv == 1.0.1
pytest == 8.1.1
telebot == 0.0.5
pyyaml == 6.0.1
bs4 == 0.0.2
lxml == 5.2.1